Refactor and clean up Terms and Conditions page and modal #723
