+++ client/src/App.js
@@ -8,7 +8,7 @@ import ProjectsPage from "./components/Projects/ProjectsPage";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import About from "./components/About";
-import TermsAndConditions from "./components/TermsAndConditions";
+import TermsAndConditionsPage from "./components/TermsAndConditions/TermsAndConditionsPage";
 import PrivacyPolicy from "./components/PrivacyPolicy";
 import ContactUs from "./components/ContactUs";
 import Register from "./components/Authorization/Register";
@@ -77,7 +77,7 @@ const App = ({
           render={() => <ProjectsPage account={account} />}
         />
         <Route path="/about" component={About} />
-        <Route path="/termsandconditions" component={TermsAndConditions} />
+        <Route path="/termsandconditions" component={TermsAndConditionsPage} />
         <Route path="/privacypolicy" component={PrivacyPolicy} />
         <Route path="/register/:email?" component={Register} />
         <Route path="/confirm/:token" component={ConfirmEmail} />
