--- client/src/components/TermsAndConditionsModal.test.js
@@ -1,15 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import Modal from "react-modal";
-import { BrowserRouter as Router } from "react-router-dom";
-
-it("renders without crashing", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(
-    <Router>
-      <Modal />
-    </Router>,
-    div
-  );
-  expect(div).toBeDefined();
-});
